plugin deps, eslint flowtype plugin
