+++ package.json
@@ -41,6 +41,7 @@
     "hapi": "16.4.0",
     "hapi-pino": "1.6.0",
     "idb-keyval": "2.3.0",
+    "inert": "4.2.0",
     "intl": "1.2.5",
     "joi": "10.5.2",
     "js-cookie": "2.1.4",
@@ -68,7 +69,7 @@
     "babel-preset-react": "6.24.1",
     "babel-preset-stage-2": "6.24.1",
     "coveralls": "2.13.1",
-    "eslint": "3.19.0",
+    "eslint": "4.0.0",
     "eslint-config-prettier": "2.1.1",
     "eslint-plugin-flowtype": "2.34.0",
     "eslint-plugin-react": "7.0.1",
